Rename file app-footer.vue to AppFooter.vue
