+++ layouts/default.vue
@@ -76,7 +76,7 @@ h6 {
 }
 </style>
 <script>
-import AppFooter from '~/components/app-footer'
+import AppFooter from '@/components/AppFooter'
 import HeaderNav from '~/components/header-nav'
 import HotNewsBanner from '~/components/hot-news-banner'
 export default {
